Split out fetch metrics fields conversion
To have smaller, more focused source files.

Bug: 549999548
