+++ base/cvd/cuttlefish/host/libs/metrics/BUILD.bazel
@@ -43,6 +43,23 @@ cf_cc_library(
 )
 
 cf_cc_library(
+    name = "fetch_conversion",
+    srcs = ["fetch_conversion.cc"],
+    hdrs = ["fetch_conversion.h"],
+    deps = [
+        "//cuttlefish/host/commands/cvd/fetch:builds",
+        "//cuttlefish/host/libs/metrics:fetch_metrics",
+        "//cuttlefish/host/libs/web:android_build",
+        "//external_proto:cf_fetch_build_cc_proto",
+        "//external_proto:cf_fetch_builds_cc_proto",
+        "//external_proto:cf_fetch_complete_cc_proto",
+        "//external_proto:cf_fetch_failure_cc_proto",
+        "//external_proto:cf_fetch_start_cc_proto",
+        "//external_proto:cf_metrics_event_v2_cc_proto",
+    ],
+)
+
+cf_cc_library(
     name = "fetch_metrics",
     srcs = ["fetch_metrics.cc"],
     hdrs = ["fetch_metrics.h"],
@@ -147,22 +164,16 @@ cf_cc_library(
     hdrs = ["metrics_conversion.h"],
     deps = [
         "//cuttlefish/common/libs/utils:host_info",
-        "//cuttlefish/host/commands/cvd/fetch:builds",
         "//cuttlefish/host/libs/config:data_image_policy",
         "//cuttlefish/host/libs/config:gpu_mode",
         "//cuttlefish/host/libs/metrics:device_event_type",
+        "//cuttlefish/host/libs/metrics:fetch_conversion",
         "//cuttlefish/host/libs/metrics:fetch_metrics",
         "//cuttlefish/host/libs/metrics:gce_environment",
         "//cuttlefish/host/libs/metrics:github_environment",
         "//cuttlefish/host/libs/metrics:guest_metrics",
         "//cuttlefish/host/libs/metrics:host_metrics",
         "//cuttlefish/host/libs/metrics:invoker",
-        "//cuttlefish/host/libs/web:android_build",
-        "//external_proto:cf_fetch_build_cc_proto",
-        "//external_proto:cf_fetch_builds_cc_proto",
-        "//external_proto:cf_fetch_complete_cc_proto",
-        "//external_proto:cf_fetch_failure_cc_proto",
-        "//external_proto:cf_fetch_start_cc_proto",
         "//external_proto:cf_flags_cc_proto",
         "//external_proto:cf_gce_environment_cc_proto",
         "//external_proto:cf_github_actions_environment_cc_proto",
