+++ base/cvd/cuttlefish/host/libs/metrics/metrics_conversion.cc
@@ -24,22 +24,16 @@
 #include "google/protobuf/timestamp.pb.h"
 
 #include "cuttlefish/common/libs/utils/host_info.h"
-#include "cuttlefish/host/commands/cvd/fetch/builds.h"
 #include "cuttlefish/host/libs/config/data_image_policy.h"
 #include "cuttlefish/host/libs/config/gpu_mode.h"
 #include "cuttlefish/host/libs/metrics/device_event_type.h"
+#include "cuttlefish/host/libs/metrics/fetch_conversion.h"
 #include "cuttlefish/host/libs/metrics/fetch_metrics.h"
 #include "cuttlefish/host/libs/metrics/gce_environment.h"
 #include "cuttlefish/host/libs/metrics/github_environment.h"
 #include "cuttlefish/host/libs/metrics/guest_metrics.h"
 #include "cuttlefish/host/libs/metrics/host_metrics.h"
 #include "cuttlefish/host/libs/metrics/invoker.h"
-#include "cuttlefish/host/libs/web/android_build.h"
-#include "external_proto/cf_fetch_build.pb.h"
-#include "external_proto/cf_fetch_builds.pb.h"
-#include "external_proto/cf_fetch_complete.pb.h"
-#include "external_proto/cf_fetch_failure.pb.h"
-#include "external_proto/cf_fetch_start.pb.h"
 #include "external_proto/cf_flags.pb.h"
 #include "external_proto/cf_gce_environment.pb.h"
 #include "external_proto/cf_github_actions_environment.pb.h"
@@ -53,13 +47,6 @@ namespace {
 
 using google::protobuf::Timestamp;
 using logs::proto::wireless::android::cuttlefish::CuttlefishLogEvent;
-using logs::proto::wireless::android::cuttlefish::events::CuttlefishBuild;
-using logs::proto::wireless::android::cuttlefish::events::CuttlefishFetchBuilds;
-using logs::proto::wireless::android::cuttlefish::events::
-    CuttlefishFetchComplete;
-using logs::proto::wireless::android::cuttlefish::events::
-    CuttlefishFetchFailure;
-using logs::proto::wireless::android::cuttlefish::events::CuttlefishFetchStart;
 using logs::proto::wireless::android::cuttlefish::events::CuttlefishFlags;
 using logs::proto::wireless::android::cuttlefish::events::
     CuttlefishFlags_DataPolicy;
@@ -264,81 +251,6 @@ void PopulateCuttlefishGuest(CuttlefishGuest& guest,
       guest_metrics.flag_metrics.restart_subprocesses);
   flags.set_system_image_dir_specified(
       guest_metrics.flag_metrics.system_image_dir_specified);
-}
-
-void PopulateCuttlefishBuild(CuttlefishBuild& cf_build, const Build& build) {
-  cf_build.set_build_id(std::visit([](auto&& arg) { return arg.id; }, build));
-  cf_build.set_target(std::visit([](auto&& arg) { return arg.target; }, build));
-  cf_build.set_product(
-      std::visit([](auto&& arg) { return arg.product; }, build));
-  if (const DeviceBuild* device_build = std::get_if<DeviceBuild>(&build)) {
-    cf_build.set_branch(device_build->branch);
-  }
-}
-
-void PopulateFetchBuilds(CuttlefishFetchBuilds& fetch_builds,
-                         const Builds& builds) {
-  if (builds.default_build) {
-    CuttlefishBuild& base_build = *fetch_builds.mutable_base();
-    PopulateCuttlefishBuild(base_build, builds.default_build.value());
-  }
-  if (builds.system) {
-    CuttlefishBuild& system_build = *fetch_builds.mutable_system();
-    PopulateCuttlefishBuild(system_build, builds.system.value());
-  }
-  if (builds.kernel) {
-    CuttlefishBuild& kernel_build = *fetch_builds.mutable_kernel();
-    PopulateCuttlefishBuild(kernel_build, builds.kernel.value());
-  }
-  if (builds.boot) {
-    CuttlefishBuild& boot_build = *fetch_builds.mutable_boot();
-    PopulateCuttlefishBuild(boot_build, builds.boot.value());
-  }
-  if (builds.bootloader) {
-    CuttlefishBuild& bootloader_build = *fetch_builds.mutable_bootloader();
-    PopulateCuttlefishBuild(bootloader_build, builds.bootloader.value());
-  }
-  if (builds.android_efi_loader) {
-    CuttlefishBuild& android_efi_loader_build =
-        *fetch_builds.mutable_android_efi_loader();
-    PopulateCuttlefishBuild(android_efi_loader_build,
-                            builds.android_efi_loader.value());
-  }
-  if (builds.otatools) {
-    CuttlefishBuild& otatools_build = *fetch_builds.mutable_otatools();
-    PopulateCuttlefishBuild(otatools_build, builds.otatools.value());
-  }
-  if (builds.test_suites) {
-    CuttlefishBuild& test_suites_build = *fetch_builds.mutable_test_suites();
-    PopulateCuttlefishBuild(test_suites_build, builds.test_suites.value());
-  }
-}
-
-struct PopulateFetchEvent {
-  MetricsEventV2& event;
-
-  void operator()(const FetchStartMetrics& start_metrics) {
-    CuttlefishFetchStart& start = *event.mutable_start();
-    start.set_enable_local_caching(start_metrics.enable_local_caching);
-    start.set_dynamic_super_image(start_metrics.dynamic_super_image_mixing);
-  }
-
-  void operator()(const FetchCompleteMetrics& complete_metrics) {
-    CuttlefishFetchComplete& complete = *event.mutable_complete();
-    complete.set_status_blocked(complete_metrics.status_blocked);
-    complete.set_fetch_size_bytes(complete_metrics.fetch_size_bytes);
-    for (const Builds& builds : complete_metrics.fetched_builds) {
-      CuttlefishFetchBuilds& fetch_builds = *complete.add_builds();
-      PopulateFetchBuilds(fetch_builds, builds);
-    }
-  }
-
-  void operator()(const FetchFailedMetrics&) { event.mutable_failure(); }
-};
-
-void PopulateCuttlefishFetch(MetricsEventV2& event,
-                             const FetchMetrics& fetch_metrics) {
-  std::visit(PopulateFetchEvent{.event = event}, fetch_metrics);
 }
 
 }  // namespace
